+++ packages/blinkcard-ux-manager/src/core/BlinkCardUxManager.ts
@@ -8,7 +8,6 @@ import {
   type BlinkCardSessionSettings,
   type DeviceInfo,
   type PingCameraInputInfoData,
-  type PingScanningConditionsData,
   type PingUxEventData,
   type ProcessResultWithBuffer,
   type RemoteScanningSession,
@@ -25,6 +24,7 @@ import {
   convertCameraToPingCamera,
   hasCameraListChanged,
 } from "@microblink/ux-common/cameraAnalyticsMappers";
+import { subscribeToDeviceOrientation } from "@microblink/ux-common/deviceOrientationAnalytics";
 import { HapticFeedbackManager } from "@microblink/ux-common/hapticFeedback";
 import { RafLoop } from "@microblink/ux-common/RafLoop";
 import { invokeCallbacks, sleep } from "@microblink/ux-common/utils";
@@ -334,44 +334,7 @@ export class BlinkCardUxManager {
 
     this.#cleanupCallbacks.add(unsubscribeCameraPermission);
 
-    // Orientation pings
-    const reportOrientation = (orientation: ScreenOrientation) => {
-      let deviceOrientation: PingScanningConditionsData["deviceOrientation"];
-
-      switch (orientation.type) {
-        case "portrait-primary":
-          deviceOrientation = "Portrait";
-          break;
-        case "portrait-secondary":
-          deviceOrientation = "PortraitUpside";
-          break;
-        case "landscape-primary":
-          deviceOrientation = "LandscapeLeft";
-          break;
-        case "landscape-secondary":
-          deviceOrientation = "LandscapeRight";
-          break;
-      }
-
-      void this.#analytics.logDeviceOrientation(deviceOrientation);
-    };
-
-    const orientationChangeHandler = (event: Event) => {
-      const target = event.target as ScreenOrientation;
-      reportOrientation(target);
-    };
-
-    screen.orientation.addEventListener("change", orientationChangeHandler);
-
-    // initial report
-    reportOrientation(screen.orientation);
-
-    this.#cleanupCallbacks.add(() => {
-      screen.orientation.removeEventListener(
-        "change",
-        orientationChangeHandler,
-      );
-    });
+    this.#cleanupCallbacks.add(this.#subscribeToDeviceOrientationAnalytics());
 
     const unsubTorch = this.cameraManager.subscribe(
       (state) => state.selectedCamera?.torchEnabled,
@@ -406,6 +369,17 @@ export class BlinkCardUxManager {
     this.#cleanupCallbacks.add(() => {
       this.stopUiUpdateLoop();
     });
+  }
+
+  #subscribeToDeviceOrientationAnalytics(): () => void {
+    return subscribeToDeviceOrientation(
+      (deviceOrientation) => {
+        void this.#analytics.logDeviceOrientation(deviceOrientation);
+      },
+      (logMessage) => {
+        void this.#analytics.logWarning(logMessage);
+      },
+    );
   }
 
   #handleCameraPermissionChange = (